add mensagens de sucesso

+++ src/pages/GerenciarUsuario.js
@@ -5,6 +5,7 @@ const GerenciarUsuario = () => {
   const [usuarios, setUsuarios] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
+  const [successMessage, setSuccessMessage] = useState(''); // Estado para mensagem de sucesso
 
   useEffect(() => {
     const fetchUsuarios = async () => {
@@ -45,34 +46,39 @@ const GerenciarUsuario = () => {
 
   // Função para deletar usuário
   const handleDeleteUser = async (userId) => {
-    try {
-      // Obter token do armazenamento local
-      const token = localStorage.getItem('token');
+    if (window.confirm('Tem certeza que deseja excluir o usuário?')) { // Adicionando confirmação
+      try {
+        // Obter token do armazenamento local
+        const token = localStorage.getItem('token');
 
-      // Verificar se o token está presente
-      if (!token) {
-        throw new Error('Token JWT não encontrado no armazenamento local.');
-      }
-
-      // Configurar o cabeçalho de autorização
-      const config = {
-        headers: {
-          Authorization: `Bearer ${token}`
+        // Verificar se o token está presente
+        if (!token) {
+          throw new Error('Token JWT não encontrado no armazenamento local.');
         }
-      };
 
-      // Enviar solicitação para deletar o usuário
-      await axios.delete(`http://localhost:8080/porkManagerApi/usuario/deleteUsuario/${userId}`, config);
+        // Configurar o cabeçalho de autorização
+        const config = {
+          headers: {
+            Authorization: `Bearer ${token}`
+          }
+        };
 
-      // Atualizar a lista de usuários após deletar o usuário com sucesso
-      setUsuarios(usuarios.filter(user => user.id !== userId));
-    } catch (error) {
-      console.error('Erro ao deletar usuário:', error);
+        // Enviar solicitação para deletar o usuário
+        await axios.delete(`http://localhost:8080/porkManagerApi/usuario/deleteUsuario/${userId}`, config);
+
+        // Atualizar a lista de usuários após deletar o usuário com sucesso
+        setUsuarios(usuarios.filter(user => user.id !== userId));
+
+        // Definir mensagem de sucesso
+        setSuccessMessage('Usuário deletado com sucesso!');
+      } catch (error) {
+        console.error('Erro ao deletar usuário:', error);
+      }
     }
   };
 
   return (
-    <div className="text-gray-900 bg-gray-200">
+    <div className="text-gray-900 bg-gray-200 min-h-screen">
       <div className="p-4 flex justify-between items-center">
         <h1 className="text-3xl mx-auto"> {/* Adicionando a classe mx-auto para centralizar horizontalmente */}
           Controle de Usuário
@@ -121,6 +127,42 @@ const GerenciarUsuario = () => {
           ))}
         </div>
       </div>
+      {successMessage && (
+        <div className="bg-green-100 border border-green-400 text-green-700 px-4 py-3 rounded relative shadow-md mx-auto mt-4 w-1/2">
+          <span className="block sm:inline">{successMessage}</span>
+          <span
+            className="absolute top-0 bottom-0 right-0 px-4 py-3"
+            onClick={() => setSuccessMessage('')}
+          >
+            <svg
+              className="fill-current h-6 w-6 text-green-500"
+              role="button"
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 20 20"
+            >
+              <path d="M14.348 5.652a1 1 0 00-1.415 0L10 8.586 7.067 5.652a1 1 0 10-1.415 1.415L8.586 10l-2.934 2.933a1 1 0 101.415 1.415L10 11.414l2.933 2.934a1 1 0 101.415-1.415L11.414 10l2.934-2.933a1 1 0 000-1.415z" />
+            </svg>
+          </span>
+        </div>
+      )}
+      {error && (
+        <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative shadow-md mx-auto mt-4 w-1/2">
+          <span className="block sm:inline">{error}</span>
+          <span
+            className="absolute top-0 bottom-0 right-0 px-4 py-3"
+            onClick={() => setError(null)}
+          >
+            <svg
+              className="fill-current h-6 w-6 text-red-500"
+              role="button"
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 20 20"
+            >
+              <path d="M14.348 5.652a1 1 0 00-1.415 0L10 8.586 7.067 5.652a1 1 0 10-1.415 1.415L8.586 10l-2.934 2.933a1 1 0 101.415 1.415L10 11.414l2.933 2.934a1 1 0 101.415-1.415L11.414 10l2.934-2.933a1 1 0 000-1.415z" />
+            </svg>
+          </span>
+        </div>
+      )}
     </div>
   );
 };
